refactor: extract page logic into dedicated module components
Move data fetching and rendering logic from parcours-conformite page.tsx
files into SecondDeclarationStepPage and CompliancePathPage components,
keeping pages as thin wrappers per convention.
